client/requirements.txt
logzero==1.7.0
pyserial==3.5
